+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [BREAKING CHANGE] Stop automatically uppercasing annotation option keys.
 - Do not mutate options passed to `doc.annotate()` and its convenience methods (link, note, strike, lineAnnotation, rectAnnotation, ellipseAnnotation, textAnnotation, fileAnnotation)
 - Persist font options when adding a new page. Fixes #1739
+- Use `Uint8Array` instead of Node's `Buffer` internally. `pdfkit/virtual-fs` now returns a `Uint8Array` from `readFileSync()`
 
 ### [v0.19.1] - 2026-06-10